requirements.txt
fastapi
uvicorn
requests
python-multipart
python-dotenv
# uvicorn main:app --reload